engine/php/requirements.txt
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
click>=8.1.0
rich>=13.7.0
pydantic>=2.7.0
jinja2>=3.1.0
PyYAML>=6.0.0
google-generativeai>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0